fix(slack): break actions/actions-bookmarks import cycle
Extract getClient and resolveToken into actions-client.ts so actions.ts
and actions-bookmarks.ts both import the shared client resolver from a
leaf module instead of through the actions.ts barrel, eliminating the
runtime value import cycle the architecture guard detected.
